Blind limb of quotient polynomials and ensure ZK (#161)
* Minor cosmetics

* Change splitting factor of quotient limbs from [x^n] to [x^{n-1}]

* Blind quotient limbs

* Adjust verifier_gadget

* CHANGELOG

---------

+++ proofs/src/poly/kzg/params.rs
@@ -43,8 +43,8 @@ impl<E: Engine + Debug> ParamsKZG<E> {
         }
 
         let n = 1 << new_k;
-        assert!(n < self.g_lagrange.len() as u32);
-        self.g.truncate(n as usize);
+        assert!(n < self.g_lagrange.len());
+        self.g.truncate(n);
         self.g_lagrange = g_to_lagrange(&self.g, new_k);
     }
